Dependency injection via constructor
In this commit we utilized the LoginComponent's
constructor to inject a new property into the
component.

The cool thing about doing this with TypeScript
is that we don't need to declare and bind this
property somewhere else to start using it.

Then, we used router.navigate to navigate from
the login page to the welcome page.