Adding and receiving a routing parameters
In this commit we passed a route parameter from login component
after a successful login.

The route into app-routing-module had to be modified as well, to
accommodate the new parameter.

After that in the welcome component, we created a variable and
assigned to it the value that comes from the route param.