Returning a Bean from RESTful API
The teaching here is to remember that DispatcherServlet plays
a key role into our app, as it is responsible for the app's
root mapping /, and from there it finds the proper resource
to handle the request.

After resource had handled the request the DispatcherServlet
receives the request back. From there, it resolves how this
response should be returned to the client, which in our case
it finds the Jackson2ObjectMapperBuilderCustomizerConfiguration
there is the responsible to converting the Bean to JSON format.

That's why we get the Task Bean as a JSON Object with its properties.