Creating a service to fetch backend data
In this important commit, we created a service responsible
for fetching backend data.

To do that, it makes use of an HttpClient module, and as
a new Angular Module, it had to be imported in app.module.ts,
and declared as a dependency into the service.

The task service making use of the http client module subscribe
to an observable, which in turn fires the request to the backend.

On it's turn, the web client component declares the service as
a dependency and the desired action.