Creating a header for requests
In this commit, we created a service that creates a basic
authentication header to be put into requests.

For now, the header is with hardcoded credentials.

As we enabled spring security basic and form authentication
with the adding of spring security to the project, requests
to the backend resource wasn't being accepted with the error
No Access-Control-Allow-Origin header is present on the
requested resource.