Connecting login page to basic authentication
In this commit we connected the login page to the
basic authentication.

In this proccess, we disabled temporarily the http
interceptor provider to not override our submitted
request credentials with those provided by the
interceptor.

And in regard to the basic authentication, we created
a method in angular that subscribes to a service that
by it's turn asks for the RESTful service for a login
and then receives a response. That response could be
some data or an error.